Migrated from double to float, added vcc voltage sensor

+++ include/Sensors/HumiditySensor.h
@@ -7,7 +7,7 @@ namespace ACC { namespace Sensors {
      */
     class HumiditySensor {
         public:
-            virtual double measureHumidity() = 0;
+            virtual float measureHumidity() = 0;
     };
 } }
